shop and booking merge

+++ src/main/java/com/cndsalon/domain/book/BookingView.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class BookingView {
 	@Id
-	private int bCode;
+	private Long bCode;
 	private String id;
 	private int bPrice;
 	private String bStatus;
